requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
discord==1.7.3
discord-together==1.2.3
discord.py==1.7.3
DiscordUtils==1.3.4
idna==3.3
multidict==6.0.2
numpy==1.22.2
opencv-python==4.5.5.62
python-dotenv==0.19.2
requests==2.27.1
speedtest-cli==2.1.3
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
youtube-dl==2021.12.17
